+++ Content.Server/Inventory/ServerInventorySystem.cs
@@ -1,15 +1,24 @@
+using Content.Server.Radio.EntitySystems;
+using Content.Shared.Access.Components;
+using Content.Shared.Access.Systems;
 using Content.Shared.Explosion;
 using Content.Shared.Inventory;
+using System.Globalization;
 
 namespace Content.Server.Inventory
 {
     public sealed class ServerInventorySystem : InventorySystem
     {
+
+        [Dependency] private readonly SharedIdCardSystem _sharedIdCard = default!; // SS220 Borgs-Id-fix
+
         public override void Initialize()
         {
             base.Initialize();
 
             SubscribeLocalEvent<InventoryComponent, BeforeExplodeEvent>(OnExploded);
+
+            SubscribeLocalEvent<InventoryComponent, GetInsteadIdCardNameEvent>(OnGetIdCardName); // SS220 Borgs-Id-fix
         }
 
         private void OnExploded(Entity<InventoryComponent> ent, ref BeforeExplodeEvent args)
@@ -35,5 +44,28 @@ namespace Content.Server.Inventory
                     TryEquip(target, item, slot.Name , true, true, inventory: target.Comp, triggerHandContact: true);
             }
         }
+
+        // SS220 Borgs-Id-fix start
+        private void OnGetIdCardName(EntityUid uid, InventoryComponent invent, ref GetInsteadIdCardNameEvent args)
+        {
+            var idCard = new Entity<IdCardComponent>();
+            var textInfo = CultureInfo.CurrentCulture.TextInfo;
+            string idCardName = Loc.GetString("chat-radio-no-id");
+            idCardName = textInfo.ToTitleCase(idCardName);
+            // Проверка слота
+            if (TryGetSlotEntity(uid, "id", out var idUid))
+            {
+                _sharedIdCard.TryGetIdCard(idUid.Value, out idCard);
+
+                idCardName = idCard.Comp.LocalizedJobTitle ?? idCardName;
+                idCardName = textInfo.ToTitleCase(idCardName);
+                args.Name = $"\\[{idCardName}\\] ";
+                return;
+            }
+            // если карту не нашли, отправляем "[Без ID]"
+            else
+                args.Name = $"\\[{idCardName}\\] ";
+        }
+        // SS220 Borgs-Id-fix end
     }
 }
