+++ Resources/Locale/ru-RU/ss220/borgs/borgs_type.ftl
@@ -1,0 +1,6 @@
+borg-type-prototype-generic = Киборг
+borg-type-prototype-engineering = Киборг-инженер
+borg-type-prototype-mining = Киборг-шахтёр
+borg-type-prototype-janitor = Киборг-уборщик
+borg-type-prototype-medical = Киборг-доктор
+borg-type-prototype-service = Киборг-официант
