+++ Resources/Prototypes/borg_types.yml
@@ -1,6 +1,7 @@
 # Generic borg
 - type: borgType
   id: generic
+  name: borg-type-prototype-generic # SS220 Borgs-Id-fix
 
   # Description
   dummyPrototype: BorgChassisGeneric
@@ -33,6 +34,7 @@
 # Engineering borg
 - type: borgType
   id: engineering
+  name: borg-type-prototype-engineering # SS220 Borgs-Id-fix
 
   # Description
   dummyPrototype: BorgChassisEngineer
@@ -69,6 +71,7 @@
 # Salvage borg
 - type: borgType
   id: mining
+  name: borg-type-prototype-mining # SS220 Borgs-Id-fix
 
   # Description
   dummyPrototype: BorgChassisMining
@@ -106,6 +109,7 @@
 # Janitor borg
 - type: borgType
   id: janitor
+  name: borg-type-prototype-janitor # SS220 Borgs-Id-fix
 
   # Description
   dummyPrototype: BorgChassisJanitor
@@ -141,6 +145,7 @@
 # Medical borg
 - type: borgType
   id: medical
+  name: borg-type-prototype-medical # SS220 Borgs-Id-fix
 
   # Description
   dummyPrototype: BorgChassisMedical
@@ -188,6 +193,7 @@
 # Service borg
 - type: borgType
   id: service
+  name: borg-type-prototype-service # SS220 Borgs-Id-fix
 
   # Description
   dummyPrototype: BorgChassisService
